id 타입 정리 및 기술 데이터 수집 설계 시작

+++ PowikiBack/src/main/java/com/example/powiki/domain/mechanic/model/entity/TypeEfficacy.java
@@ -15,8 +15,8 @@ import java.time.LocalDateTime;
 @ToString
 public class TypeEfficacy {
 
-    private Integer damageTypeId;
-    private Integer targetTypeId;
+    private Long damageTypeId;
+    private Long targetTypeId;
     private double damageFactor;
 
     private char isActive;
